♻️ Refactor large components into smaller, focused modules
- Split SlotMachineWheel into smaller components (wheel/WheelSegment, wheel/WheelFrame)
- Extracted spin animation logic into useSpinAnimation hook
- Split OptionsPage by extracting CSV import logic to useCSVImport hook
- Created CompetitionManagement component for better separation
- Removed unused SpinnerWheel.tsx component
- Reduced file sizes: SlotMachineWheel 310→133 lines, OptionsPage 274→146 lines

This improves code maintainability and follows single responsibility principle.
